Bug#35012146 Multivalued index creation results in partial  result for query with ORDER BY DESC
The descending order ref access handler has a logic for comparing the
result row against the key value to avoid out-of-bound records. This
filtering is not suitable for multi-valued indexes due to the nature
of key value and row data. This generic binary comparison of fetched
column against the key from MEMBER OF operator fails and causes the
partial result.

During multi-valued index processing, it is possible to fetch
duplicates as described in wl#8955. A unique value filter logic is
present in the handler layer for most code paths including ascending
order ref index reads. This is missing in case of descending order
ref index access causing duplicate rows after access check fix.

Solution: Implemented MEMBER OF comparison instead key memcmp with
the fetched row in the handler layer for multi-valued indexes. And enabled
unique value filtering for descending ref index access path in handler.

Change-Id: Ic9fde0ad9af83848b761f916a279d0b86449f3d6