Bug#35410465: Subquery ORDER BY is ignored
This is a regression from Bug#34890862. That fix
made a change for already resolved ORDER BY items
to not be re-resolved (it is usually the case when
a derived table is merged). As part of the change,
when comparing the order by items to the visible
fields in the outer query block, it compared the
order by item with the referenced item in the
ref_item_array. For the case in bugpage, the order
by item itself is a reference. This makes the
comparison fail.
So we now compare the real_item() of the order by
expression to the real_item() of the ref_item_array
elements.
Also, when a hidden item is added, the original fix
missed to set order->item to point to the correct
ref array item like what is done in
find_order_in_list().

Change-Id: I00786757ef64df7390ad20c971baa7086d214e18