Bug#35389533 Backup won't start with NOWAIT option after restarting a data node
Problem:
A backup will not start if NOWAIT option is specified after a
data node is restarted.

Analysis:
1- if the restarted node is the master, master will eventually
change and start backup signals will end up in a wrong data
node(previous master)
2- if the backup is launched with nowait flag, the Management
Server will not wait for any reply from the backup block.

The way backup block has to report that self it not the master
is to send back a GSN_BACKUP_REF signal to Management Server.
So when both 1 and 2 are met, as Management Server is not
waiting for signals from backup the REF will be ignored and
consequently Management Server will not (re)try to perform the
start backup.

Solution:
From Management Server always wait for BACKUP_REF/BACKUP_CONF from
backup (even if nowait is specified). From backup side always send 1
signal to Management Server, GSN_BACKUP_REF if self is not master or
signal GSN_BACKUP_CONF otherwise.

Change-Id: Ie9c9163edf5b980676e72651f92cd73368d87fa9