WL# 13968 : Deprecate sync_relay_log_info
This worklog deprecates system variable sync_relay_log_info
without alternative. This option is a no-op when storing
replication metadata in system tables or in InnoDB undo log.
This patch adds the warnings and help texts, and changes the
test cases result files to accommodate the new deprecation warning
emitted when the system variable sync_relay_log_info are tweaked.
The new warning is recorded in the result files.
This patch also contains the code to trigger warning when
--sync-relay-log-info is set through command line or cnf file.

Change-Id: I3a6bdb5f282fbe4ca2001774ac99292872805b60