Bug#35362545 Fix the innodb.ibd2sdi test failing on Windows platform
Description:
------------
This test fails on some Windows machines. The test generates the
certificates in the datadir using a deprecated utility
mysql_ssl_rsa_setup. This utillity uses installed openssl to generate
the certificates. Some Windows machines may not have real openssl
version but a simulated openssl version. The simulated version generates
the certificate with different length than the one which is recorded.
Hence the test fails due to result content mismatch error.

Fix:
----
There is no need to generate the *.pem file every time test is executed
as the intention of the test is to verify the negative cases by feeding
these certificates to the ibd2sdi util. We can test the same by
pre-generated certs. Created a std_data/innodb/data dir and copied the
required certificates which are genreated in a datadir.
Used copied certificates from to verify the tests scenarios.

Change-Id: Ia0e7d77e492eb2eea4ce8f0fc8dc44b76f3d4d3b