Bug#35197060 Remove the NO_FALLOCATE macro from the code
Description
-----------
NO_FALLOCATE macro was introduce through Bug#18069105 fix. This macro
is not defined in the cmake but looks like it assumes that macro may
be defined in the Linux Kernel. The current Linux versions that MySQL
supports do not have this macro defined. That means these Linux
versions supports the fallocate() call. That means this macro serves
no useful purpose in the code. This could be removed safely to reduce
maintenance hassles.

Fix
----
Remove NO_FALLOCATE macro from the innodb code base.

  Change-Id: Iaf5b78ef1ae4b641058a211f0a93c65f08r96a6

Change-Id: I1421ef55d55e8baa34e966ac505946f698182508