BUG#35424845 Reset custom data when dropping event op
Problem:
The ndbcluster plugin subscribes to row changes in NDB using
NdbEventOperation's and as part of this is stores a pointer to
application specific data structure in "custom data". When dropping a
NdbEventOperation the application specific data structure is released but
the pointer in "custom data" is not reset. Normally the dropped
NdbEventOperation will never again be returned from any of the NdbAPi
event iterator functions but in case it would, the stale pointer would
then be used.

Analysis:
Potentially using a stale pointer leads to problem which are hard to
diagnose, it's better to have consistent behaviour that immediately
indicates the source of the problem.

Solution:
Reset the "custom data" pointer as part of dropping the
NdbEventOperation and the applciation specific data.

Change-Id: I4ecc9382f2de85074e68b1aabf6a6f568442f508