Bug#35240913 Assertion `path->cost >= 0.0' failed.
In EstimateMaterializeCost(), we calculate the cost of 'MATERIALIZE'
paths by adding the cost of materialization to the sum of the cost of
the child path(s).  If the number of output rows is undefined for a
child, we ignore that child, as we assume that the cost of that child
is then also undefined. But if the child is an 'AGGREGATE' path with
implicit grouping, the number of output rows may be set (to 1), even if
the cost is undefined. This is what triggered the assert.

The fix is to check if the cost of the child is defined, and to skip it if not.

Change-Id: Ibadfec0505ea23542dd2789428a5ddf1eaf09c09