Bug#35194724 Fix the incorrect usage of the count_sessions.inc in the innodb suites
Description:
------------
Usually the tests that use multiple connections include the script
count_session.inc and its counterpart wait_until_count_sessions.inc
to keep track of open connetions.  It was found that some of the tests
were using these scripts incorrectly.

Fix:
----
This patch fixes the wrong usage of the scripts found in some of the
tests.

Change-Id: I61c4c8df066d4183c4e0a9ed1623d56b627bdf1e