Bug#35199996 table map entry for certain tables are logged in binlog
Problem:

On a distributed setup, where on each site every mysqld replicates
changes from each other but are set to not log these updates on their
own binlog, a Table_map entry was eventually logged. This causes the
replication applier of a 3rd server to fail with ER_NO_SUCH_TABLE.

Analysis:

1. The Table_map events belonging to changes that were applied by a
replica were being injected without looking at the log_replica_updates
option. The code was only guarding against the set of events within the
epoch that had the NDB_ANYVALUE_NOLOGGING_CODE flag set in the
anyValue. However, this flag is only set for the specific cases that
binlog is disabled.

2. The cumulative_any_value of an event operation, a bitwise AND
operation over the any-value of all events for a given table belonging
to the same epoch, was not being used correctly for a number of
reasons: i) It was being relied on to give an overview of who had made
changes on the table over the course of the epoch; ii) It could
contain reserved values, whose bits are the inverse of the bits used
to record server ids; iii) It was being reduced to a bitwise AND of
all recorded values on each event, thus causing cancellation between
inverse bits.

Solution:

1. The injection of a Table_map entry for changes occured in the table
is now guarded with the log_replica_updates option.

2. The cumulative_any_value is now created with a bitwise OR operation
so that information is not lost. Additionally, it is filtered by the
plugin event operation on every processing of the event data so that
important data is retrieved, such as if the update was made by a
replica or if it had a no-logging flag set. This data is then used for
a correct decision on whether to inject a Table_map event.

Changed rpl_conflict tests' setup since mysqld.1.1 must log replica
updates and the default is turning it off.

Change-Id: I978325fe01435a7d0e0bd4249979b749731f8fbb