Bug#35419418 CH Benchmark failing with crash in histogram code
When building a histogram on a column that contains string data, the
strings that the histogram points to are allocated on a mem_root that
only lives within the scope of the function that builds the histograms.

This results in a reading-freed-memory type of failure when we later,
within the same transaction but in a different scope, attempt to create
a copy of the histogram to be placed on the table_share, thereby reading
from memory that has already been freed when copying the strings.

This failure only shows up for string histograms since other histogram
types do not suffer from the problem of heap-allocated data on a
mem_root that goes out of scope.

The failure also only happens when updating an existing histogram. The
first time we create a histogram on a colum it is written directly to
the dictionary tables, skipping the shared dictionary cache and local
object registry. When later in the same transaction we retrieve the
histogram from the dictionary we get a newly created object directly
from the dictionary tables with strings that are allocated on the
histogram's own mem_root. However, when we update an existing histogram
it is stored in the dictionary client's local object registry, and when
we later acquire the same object and attempt to copy it, the mem_root
that the strings were allocated on has gone out of scope.

This patch fixes the above failure by ensuring that strings are copied
onto the histogram's own mem_root when we build a new histogram,
bringing the behavior of histograms::build_histogram() in line with its
documentation.

Change-Id: Ic9d669dc9012d25a3442c48e5423e02118135e31