WL#15294 Extending GTID with tags to identify group of transactions - step 4
Step 4. Changes in GTID handling/generation in the GR plugin

This step performs integration of the GTID tag feature with the Group
Replication plugin. It extends algorithms for GTID generation
used in the Ceritifier class.

* Background:

Before this WL, Certifier kept track of:
- GTID free intervals for the group UUID
- currently "allocated" GTID interval used for each member.
Certifier allocated GTID intervals in blocks, with specified block size.
In case the GR view changes, certifier is updating the list of available
GTID intervals. Allocated blocks are cleared out, list of available
intervals is re-calculated from the current state of the group gtid
executed set, then new blocks are allocated. In case the current
GTID interval is exhausted, a new block is allocated for the group sidno
and member from which transaction originates.

* Overview of changes:

In this step, functionality is extended to handle all TSIDs
which correspond to different SIDNOs (tagged TSIDs with the group
replication UUID). Descibed logic is refactored out to the newly
introduced "Gtid_generator" and "Gtid_generator_for_sidno" classes.
Gtid_generator keeps track of generators of GTIDs for recorded
sidnos (AUTOMATIC tagged or untagged).

Cerification handler utilizes new information about GTID to be generated
(tag) and propagated in the updated Gtid_log_event.

Change-Id: I7d0a7fc4df975d8772ac97c444216623db5b98d6