Bug#35442407 Server exit when executing query
The server now exits where we used to see an error message.  The
refactoring performed in commit:

  WL#14672: Enable the hypergraph optimizer for UPDATE [7/8, single-table]

has (presumably inadvertently) inverted the logic when we check for
the presence of ORDER BY and LIMIT with multi-table updates. It seems
clear that we should give an error message if ORDER BY or LIMIT is
combined with a de facto multi table update, whether it syntactically
looks like one, or it is hidden inside a view definition.

The solution is to invert the test back, so that these checks are
performed.

The fix solves Bug#35438510.

Change-Id: I662775324cc9738e283d298989572077ac8a696d