Bug#34758905 Range analysis for IN list has O(n^2) runtime complexity
During optimization, range SEL_TREE creation uses different logic
based on the LHS of the IN predicate. For a field item, each RHS
value is added to an OR tree to create the necessary expression. In
the case of a row item comparison, a DNF expression is needed. A DNF
expression is created by adding an AND tree with column values to an
OR  tree for each set of RHS values. Instead of adding each new AND
tree to the OR tree, the OR tree is added to the AND tree causing
exponential time during the tree merge.

Solution: Change the tree joining the order. Instead of adding OR
tree to the AND tree, add AND tree to the OR tree.

Change-Id: I93503f3a8e3d6fd181aa40364d24f2868f294168