Bug#34763224 Hypergraph orders and-terms inefficiently in subquery
If a (top level) predicate is a conjunction of Boolean terms, the
Hypergraph optimizer orders the terms according to selectivity, so
that the most selective terms are evaluated first.

This ignores the fact that the cost of evaluating terms may vary.
Reading a string field (or any variable-size field) is more expensive
than reading a fixed-size field (such as an integer). And the more
fields that must be read, the higher the cost.

(Note that Hypergraph *does* evaluate terms with subqueries or calls
to store procedures or user-defined functions last, regardless of
their selectivity.)

This commit calculates a cost figure for each term, based on the
number and types of fields that are read. The terms are then ordered
in the way that should give the least total cost.

Change-Id: I6cb8af97c30870d36bd8b0fb283dbd954ca735e1