Bug#35170671 mysqld crash - Assertion failed: `table_num_to_node_num[table_num] != -1'
In this repro, we attempt to transform a scalar subquery to derived
table.  In preparation for that, we see the the top level query is
implicitly grouped, so we move the grouping into a first derived
table, as per the logic in Query_block::transform_grouped_to_derived,
as a preliminary step.  In the end, we do not transform the scalar
subquery to a derived table, because it requires an unsupported
transform: it is correlated two levels deep. But the initial transform
in transform_grouped_to_derived is still carried out. This is ok, and
should result in a valid query nonetheless. However, there is a bug in
that transform: Item_subselect::replace_item neglects to look at join
conditions in subqueries when substituting references fields.
It also didn't descend into non-derived table subqueries.

The fix plugs those holes.

Change-Id: I592db84192501c55db8f925b217c7b0625347b4d