Bug#35448784 Router randomly crashes on exit if tls sessions resumption and con pool enabled
WL#15573 "TLS resumption - tls-session-cache" added registering SSL callbacks,
responsible by the session cache management to the TlsClientContext class.
Those callbacks use the TlsClientContext object members, which means they are
not allowed to trigger once the object is already destructed as this leads to
a crash.
The TlsCLientContext destructor unregisteres the callbacks to avoid that,
but this still leaves a small race-condition window for the issue.
One thread can start executing the callback, the other in the meantime can
destroy the object when the Router stops - when the first thread execution
resumes, this leads to a crash.

This patch moves the TlsClientContext out of the routing plugin start local
scope to the global scope, and postpones destructing of those objects to
the plugin deinit() function.

Change-Id: I34e6da430a03ddff66be28ae33c71d43e839b618