Bug#35302490 trx_purge() really doesn't need goto
ulint trx_purge(ulint n_purge_threads, ulint batch_size, bool
truncate) is a function used by purge coordinator to split the next
batch_size records between `n_purge_threads-1` server workers, and
itself, and do the purge (and then truncation).  The function has an
`if` which hahndles `n_purge_threads>1` as a special case, where the
body of `if` has a `goto` to jump to the body of `else`, looking
unnecessarily complicated.

Change-Id: I3cdf508159739d90b505acbdb60e732ca7e5126d