Bug#35415915 	InnoDB: log0log_log_random_disturb_Test crashes
The test crashes on division by 0, which was executed when we try to calculate recovery pages per instance required.
The calculations now allow for having mocked values when `log_test` is not `nullptr` and we don't want to couple the unit test to the Buffer Pool, as we don't have any other means of mocking it.
Additionally, the test was not deinitializing the InnoDB modules it was initializing, which crashes on Debug runs, and it is fixed now.

Change-Id: Icd0c50dd463e4b0bc8516a7b2f3446c5ccbf9ab9