Bug#35453506 - If two router services are created on the same host, then one crashes with Exception code: 0xc0000005
When the Router was executed with service related parameter
("--service", etc.) it would pass the local variable to the arg_handler
assuming it would not be used in the handler. That was not the case,
even though the parameter was handled properly elsewhere, it is not
removed from the main handler and it accesses the released memory.

This patch uses the pointer in place of the reference and passes
nullptr to the main handler which is handled properly.

Change-Id: I26c7782551bfc42dfdb40ee074513ab7d5c780eb