Bug#35102220: Assertion failure in               Item::walk_helper_thunk at ../sql/item.h
When a condition that needs to be pushed down to derived
table has one outer referenced field and another local
to the query block, assertion which checks for consistent
"depended_from" information for referenced items does not
hold good leading to the failure. This is because, after
collecting information related to a field in an expression,
"depended_from" and "context" is not cleared before
collecting information for the next field in the expression.
This leads to checking "depended_from" of an outer reference
against the field which is not leading to the assertion
failure.
Clear the "depended_from" and "context" information before
looking into a new field in the expression.

Change-Id: I052c6f48864c3109420ab9e9ee2ebae52fc22492