Bug#35453048: mysqld 8.1+ needs its own sections in files for messages to clients/to error-log
A new range of error-numbers must be started for mysqld 8.1+ before it becomes available.
This is necessary because error messages / symbols that exist in multiple server versions must have the same error number in each server version.
If we do not wish to distribute each server version with the message set of the latest version (i.e. the superset of all messages, aka the superset of messages used by all servers), it is necessary to allot some space where the 8.0 message set can still grow (in case of backports, etc.). This is achieved by placing appropriate directives to start a new range in the files messages_to_clients.txt and messages_to_error_log.txt. Certain .result files for 8.1 tests then need to be updated to reflect the fixed error-numbers. The starting index of each range has been agreed between the various stakeholders.

Messages to clients:
- start a new range of error-numbers for 8.1+ at 6,000
- 8.0 messages (for backports etc.) still go to the range starting at 3,500 (currently at ca. 4,200)
- move router range (introduced in 8.1) to 4,500
- adjust .result files for existing 8.1 tests where necessary

Messages to error-log:
- start a new range of error-numbers for 8.1+ at 15,000
- 8.0 messages (for backports etc.) still go to the range starting at 10,000 (currently at ca. 14,000)
- adjust .result files for existing 8.1 tests where necessary

Change-Id: Ib624cdbda97ffa64ec5c12a3a266ef255e710cf1