Bug#35392640: Group Replication primary with replica blocked by view change
In a GR setup, if a source of transactions exist besides the applier
channel then the following can happen:

 - There are several transactions being applied locally, already
   certified and so associated to a ticket, lets say ticket 2.
   These transactions did not yet commit.
   Note that these can be local transaction or come from an async
   channel for example

 - A view happens that has ticket 3 and that has to wait on
   transactions from ticket 2.

 - The view change (VC1) enters the GR applier channel applier and
   gets stuck there waiting for the ticket change to 3.

 - Now there is another group change, and another view change (VC2)
   while the transactions from ticket 2 end their execution.

Issue: There is a window where the last transaction from ticket 2
already marked itself as being executed but before popping the ticket,
VC2 will pop the ticket instead but never notify any of the
participants.

VC1 stays waiting for the ticket to change forever and the worker
can't be killed.

Solution:

Make the condition wait to break in periods of 1 second so the loop is
responsive to changes to the loop condition.
We also register a stage so the loop is more reponsive to kill signals.

Change-Id: I86eb6d1e470d9728c540f2fbcfb4ba9357eba103