Bug#35467047 connection sharing disabled after cache-user over plaintext channel
If the user sends a change-user, the server switches the authentication
method and initiates the authentication flow.

If the client side connection is secure (SSL or unix-socket), the client
either sends the plaintext password for caching-sha2-password, or router
can request it.

If the client side connection is plaintext, router currently forwards
the packets as is which means it is not getting the password.

Change
------

If the client-side connection is plaintext and change-user between
router and server switches to mysql_native_password, use
caching_sha2_password between router and client.

If the client sends a scrambled password (fast-auth), request the full
auth.

- Extended the change-user tests to check if connection-sharing is
  working after change-user.

Change-Id: I205d8ce9a51d5827877d02e9cf68340f30a1e2dd