Bug#35395965 heap-use-after-free with functions, procedure            and Item_func_in::populate_bisection
As part of a performance improvement, m_const_array is introduced to
save IN list items across execution. This list is populated either at
the end of preparation or at the start of execution based on the
nature of its objects. Some of its elements (String class object)
have memory allocated from runtime memory (example: input constants)
that are not valid after the end of the current execution. Ownership
of such memory is not passed to the String object at the assignment.
But reuse is attempted in subsequent runs that can lead to unexpected
behavior.

Fix: Release all runtime memory and cleanup pointers in m_const_array
at the end of statement execution.

Change-Id: Ib56b047c54b1b0d9428e9ed515e097ab4e0d5e4d