Bug #23036096 - INNODB LAST DETECTED DEADLOCK NOT SHOWING FULL QUERY
The last detected deadlock section of the engine status log is only showing
1024 characters (combined) of the thread information and the query.

Fixing the issue by removing the limit on printing the query string.
Reasons for this fix:
1. The thd_security_context() expects the user to give a sufficently
large bufffer to write.
2. It is difficult to guess the right size of the buffer without
creating a dependency on internal format of the message created by
thd_security_context.
3. Refactoring thd_security_context is not a good idea as it seems to
be a part of plugin's API spec, instead we are just truncating the
message created by thd_security_context.
4. Instead of trimming the query, we set the length of the buffer to
max_query_len + 1 KB (so as to hold any additional information)

Change-Id: I97d2f5afae9cfb008990b69686508e312c02c7cf