Bug#35168785  Hypergraph optimizer lacks separate cost constants
The user-configurable cost constants that are exposed through the tables
mysql.server_cost and mysql.engine_cost are shared between the original
optimizer and the hypergraph optimizer.

As part of building a new cost model for the hypergraph optimizer we
would like to be able to modify the cost constants used by the
hypergraph optimizer without affecting the original optimizer. For now
it suffices to use a separate set of hard-coded constants for the
hypergraph optimizer, but eventually some of the hypergraph cost
constants should also be user-configurable.

This patch adds separate hard-coded cost constants for the hypergraph
optimizer. A test verifies that changes to the user-configurable cost
constants results in plan cost changes under the original optimizer
while leaving the plan cost for the hypergraph optimizer unchanged.

This patch also fixes a potential race condition in the cost constant
cache.

Change-Id: I9bf1840af7c0d68e80ad3b8b8c2dc7e97b17d4f3