Bug#34900334: Bug that can make MySQL fail
If order by references a subquery in select list through
an alias, and if the query block having the order by gets
merged, then while deleting the unused merged derived
columns, the subquery gets deleted.
This is not correct as it is still being used in order by.
To determine if a merged derived column is used or not,
"derived_used" is checked. "derived_used" is not set for
this case because resolving for order by does not make a
call to fix_fields() when it is resolved against alias in
the select list.
Solution is to set derived_used for the order by
expression when merging the derived table's order by into
the outer query block.

Change-Id: I1bfcc25ea7493fe6b5db9ea27259e8898c0a85c4