Bug#35211828 DERIVED CONDITION PUSHDOWN WITH ROLLUP GIVES WRONG RESULTS
For a query which has a derived condition pushdown where a column
in the condition needs to be replaced, the search for a matching item
fails even though the item is present. This occurs when the
replacement item is wrapped in a rollup wrapper while the matching
item is not. The matching item does not have a rollup wrapper
because the fix for Bug#33349994 causes rollup items to be unwrapped
before comparison.

This is fixed in Item_func comparison by checking if the searched
item is wrapped in a rollup wrapper, and using the inner item for
the comparison. The item-search is modified to use the Item_func
comparison instead of explicitly unwrapping rollup items.

Change-Id: I8e148dc4e7651dcbd0c40c84ec36c03fe0267aa9