Bug#35404584 Multi-column indexes with writest will cause HA_ERR_FOUND_DUPP_KEY
Issue description
-----------------
Replication uses write-sets to detect dependencies and/or conflicts among
transactions. The write-set is extracted from primary, not null unique and
not null foreign keys when the binary log format is ROW and the
option`transaction_write_set_extraction` is enabled, both options are enabled
by default.

For string columns, indexes can be created that use only the leading part of
column values, using col_name(length) syntax to specify an index prefix
length.

Analysis
--------
It was discovered that multi-column unique key write-sets were not properly
generated, more precisely it was considering that unique keys could only
refer the leading part of column values, while the whole values of referenced
key columns are added to writeset.

Proposed solution
-----------------
In add_pke(), the values of the columns related to the multi-column key
should not be added to the writeset without truncation. Only the specified
width should be added.

Change-Id: I6d22b2e703096feb315bb85c50e6f2768ff25380