Bug#35420131  NDB: Restore will not update 'equal' PK values changed during backup
If the same PK attr is specified twice in LE_UPDATE log entry, the
second attr value will be an update of that attr value: We set up the
backup triggers to only contain attrubutes having a change in
their value.

Currently restore will just ignore PK attribute values already seen
when restoring log entries. Thus such changes will be lost.

Note that an LE_UPDATE to a PK attribute will only be generated if
we update the PK value to a 'compare-as-equal' value. E.g. update
'xyz' -> 'XYZ'. We still want such updates to be included in a restore though.

If the PK update is to a 'not-compared-as-equal' value it will be
represented as a LE_DELETE+LE_INSERT, which is not affected by
this bug.

Patch will detect that a PK-attr value is specified for the second
time in a LE_UPDATE, and generate a setValue() operation for that.

A few minor const-corrrectness fixes in addition, as well as
reducing the size of the 'keys' Bitmask which was defined to use
4096 'words' -> 128k bits.

Change-Id: I57e9b354f56bf47f19d54be98436112b2933b46c