Bug#35420131  NDB: Restore will not update 'equal' PK values changed during backup
Addendum patch, removing usage of deprecated setValue() and equal() methods taking
a 'length' argument.

These methods seems to have been marked as deprecated since:

-----------------
commit 88fbfc489dcae96f962ff2b46ebb6909f66ad710
Author: jonas@perch.ndb.mysql.com <>
Date:   Mon Nov 7 12:19:28 2005 +0100

    Import ndb varsize
-----------------

The 'length' argument is now unused. Thus using this deprecated variants
provide now extra safety check, which might have been the intention with
allowing it to be specified.

The previous patch for Bug#35420131 introduced an extra ::setValue().
As it make no sense introducing new usage of deprecated methods, we
did not use that variant of the method here.

However, there are several other places where these deprecated variants
are used in this code component. This is a good time to clean out the usage
of these as well.

Change-Id: Iee1463d9ba3315fe2055a47996e8f717afaf2001