Bug#35492532 sig 11 in decimal_round
This is a regresion from
    Bug#35380560 Item_typecast_signed::val_int(): Assertion `is_nullable()' failed.
which sets null_value to true only if an item is nullable.

Add error handling in
Item_func_min_max::decimal_op(): propagate a zero-valued object rather than nullptr.
TemptableAggregateIterator::init(): abort execution when an error occurs.

(each individual of the two changes is actually enough to fix the failure)

Change-Id: Ibaef0c383f3eb4fd601d4ae176ac0c1825f7d412