Bug#35499232 The quote function may return binary data
Before this patch:
mysql> select quote(NULL);
+--------------------------+
| quote(NULL)              |
+--------------------------+
| 0x4E554C4C               |
+--------------------------+

The fix is to never return a binary string from the quote() function,
as an 8.0 client will display the result in hex.

Documentation: Append this to documentation for function QUOTE:

The function QUOTE returns a string in the same character set and
collation as the input string, except when the input string is a
binary string. In this case, the returned string is in the connection
character set and collation.

Change-Id: I8e95d4fb06581b680933d3394d8f9ad14993f080