Bug#35497831 Assert failure - subquery_to_derived=on and a grouped correlated subquery
In debug builds we get an assert failure for the repro query during
transformation of a scalar subquery to join with derived table

Transform is only attemped for Heatwave or if the optimizer_switch has
subquery_to_derived set to ON.

This assert triggers because during transformation we create a derived
table. In this process we create a field whose type is based on
Item_long. However, the created field is of type Item_longlong, which
triggers the assert. This is expected behaviour, so the solution is to
remove the assert.

Change-Id: I87db4c8e5e09451444d45aa929fe1134f48eb246