Bug#35473657 Scalar correlated subquery with group by clause mistakenly transformed
This issue shows a case not handled by the original design for
WL#13520 Transform correlated scalar subqueries: when the scalar subquery
is grouped in addition to being correlated, a correct transformation need
to check that *for each partition of the result set, as partitioned by
the inner expression (columns) being added to the group by*, there is only
maximum one row in the derived table constructed.
This can be checked by adding using a window function

SELECT * FROM .. WHERE
      (SELECT ie FROM inner_tables WHERE inner_expr_1 = outer_expr_1 AND
                                           :
                                           inner_expr_n = outer_expr_n)
       GROUP BY e1, .. en) op oe
->

SELECT *
FROM ...
     JOIN
     ( SELECT inner_expr_1, inner_expr_n,
              COUNT(0) as aggr_cnt,
              COUNT(0) OVER (PARTITION BY inner_expr_1, ,, inner_expr_n )  AS win_cnt
       FROM inner_tables
       GROUP BY e1, .., en, inner_expr_1, .. , inner_expr_n) derived
WHERE   derived.inner_expr_1 = expr_1 AND
        :
        derived.inner_expr_n = expr_n AND
        derived.ie = oe AND
        REJECT_IF(derived.aggr_cnt > 1) AND
        REJECT_IF(derived.win_cnt > 1)

that is, even if the group by count (aggr_cnt) does only have a single row,
when we partition *on a subset* (inner_expr_1, ,, inner_expr_n) of the
expression the count may well be higher than one, and this new check
takes care of that (win_cnt).

Change-Id: I32791560c7bc84f9732d28dea887a405f52a3eb2