Bug#35493391 - Do not fail CREATE FUNCTION/PROCEDURE                if MLE component is not installed
To avoid breaking replication, we will give a warning, not error, when
creating an external routine when the component is not installed.
This means that the data dictionary may contain routines in languages
that are not supported by the server. Attempts to execute such
function will fail with an error.

Change-Id: I6e1a8b7d98e54f02aabfb31c021f624ba9c344f8