Bug#35498378 MYSQLD CRASH - ASSERTION NULLPTR != DYNAMIC_CAST<TARGET>(ARG) FAILED
While building interesting orders, the hypergraph optimizer checks if
an ordering can be used for sort-ahead only by ensuring that the
ordering is not on an expression which is in the join's field list.
This comparison incorrectly returns a match on a rollup wrapper whose
inner item matches the ordering item. Based on this incorrect
equality result, hypergraph attempts to cast the rollup wrapper item
to the ordering item type. This does not succeed because the actual
matching item is the inner item, while the cast is attempted on the
rollup wrapper.

This is fixed by removing the rollup wrapper handling from the
Item_func equality check, since it is demonstrated that unwrapping a
rollup wrapper for comparison is not appropriate in all cases. The
rollup wrapper handling is moved to the specific case which requires
it, i.e. the function to find an item in a list.

Change-Id: I9bc6df00e84868db584478c9a0b29f84deab1073