Bug#35489153 String* Item_str_func::push_packet_overflow_warning(THD*, const char*): Assertion `is_nullable()' failed
For the repeat() function we may evaluate the count argument during
resolving if it is determined to be constant.  The problem was that we
evaluated it to -42 at resolve time, and 18446744073709551574470 at
run time. This was caused by query rewriting introducing cached values.

The fix is to do count evaluation in Item_func_repeat::resolve_type()
the same way as we do for rpad()/lpad(), use val_uint() rather than
val_int().

Change-Id: I3230e0563a5ff95a840089e6f796ce9913a72921