Bug#35493484 - AddressSanitizer: heap-buffer-overflow                when starting a line with $
This is a regression from Bug#35303407

When the client encouters a dollar sign, it looks at the preceding
character to check if the dollar sign is part of an identifier.  The
fix is to skip this at the first position of the line.

The ASAN issue only occurs in interactive mode, since allocation is
different in batch mode.  Hence, the fix can not be verified by MTR
tests.

Change-Id: I3a6d9390d22d699f394c312c7b1e4bf6ca17076c