Bug#35504625: INTERSECT query assertion error
Backport of fix for Bug#35362424 in 8.0+

When pushing a condition down to derived table,
the condition to be pushed is cloned by re-parsing
the condition. To re-parse the condition, we use
Item::print to print the entire expression. While
printing an empty hex string, it prints "0x" making
the parser think that it is a "field" instead of an
hex string which results in failure later as there
is a mismatch between the number of fields in the
original expression and the cloned expression.

Modify the print function to handle empty strings
differently i.e by printing X'' instead of 0x.

Change-Id: I16f4d788c204d7fce77cc8d70217ec7ea1e4d418