Bug#35450016 'ndb_log_update_minimal=ON' is not able to replicate primary key updates
The minimal binlog format excluded all primary key columns from the
'after values' reflecting the updated row. The rational for this
was a flawed assumption about the primary key being 'constant'
when an UPDATE trigger is received.

However, if the primary key contain character data type, an UPDATE
trigger will also be received if character columns are updated to
values being 'equal' by comparison rules of the collation set being
used. E.g' 'xyz' -> XYZ'.

In order to being able to replicate such changes, we need to include
them in the after values as well.

This patch midified Ndb_event_data::init_pk_bitmap() to set up
a 'MY_BITMAP pk_nonchar_bitmap' in addition to the existing pk_bitmap.
The 'nonchar_bitmap' contain only the PK columns not being of a character
data type -> Will be 'constant' in an update trigger.

Ndb_event_data::generate_minimal_bitmap() is modified to exclude
only the nonchar_bitmap from the after values.

Note, that ideally we could (in addition) have compared the before and
after PK-values in DbtupTrigger, and included only those which changed
to a (binary) different value. This was attempted in another version
of a patch, but didn't succeed to make this work as intended.

It would also have introduced upgrade issues which would had
to be solved.

Change-Id: I90281cc6ffec1f3495f2336593547ceb7972d907